requirements.txt
# Unicrium Network - Production Requirements

# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Cryptography
pycryptodome==3.19.0
ecdsa==0.18.0

# Storage
rocksdict==0.3.20

# Networking
aiohttp==3.9.1
websockets==12.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
requests==2.31.0

# CORS
flask-cors==4.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring
prometheus-client==0.19.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
